+++ bin/cli_exit.ml
@@ -5,7 +5,9 @@ let with_defaults custom =
   let is_overridden default =
     List.exists
       (fun info ->
-        Cmdliner.Cmd.Exit.info_code info = Cmdliner.Cmd.Exit.info_code default)
+        Int.equal
+          (Cmdliner.Cmd.Exit.info_code info)
+          (Cmdliner.Cmd.Exit.info_code default))
       custom
   in
   Cmdliner.Cmd.Exit.defaults