+++ lib/properties.ml
@@ -4077,9 +4077,14 @@ let normalize_property_value : type a.
   | Display -> normalize_display value
   | Overflow -> normalize_overflow value
   | Transition -> map_preserve normalize_transition value
+  | Webkit_transition -> map_preserve normalize_transition value
+  | Moz_transition -> map_preserve normalize_transition value
+  | O_transition -> map_preserve normalize_transition value
   | List_style -> normalize_list_style value
   | Transition_timing_function -> normalize_timing_function value
   | Animation -> map_preserve normalize_animation value
+  | Webkit_animation -> map_preserve normalize_animation value
+  | Moz_animation -> map_preserve normalize_animation value
   | Animation_timing_function -> normalize_timing_function value
   | Padding_left -> Values.normalize_length ~ctx value
   | Padding_right -> Values.normalize_length ~ctx value